+++ src/modules/rules/field-date.ts
@@ -2,5 +2,6 @@ export async function updateDate(item: Zotero.Item) {
     const oldDate = item.getField("date") as string,
         newDate = Zotero.Date.strToISO(oldDate);
     newDate ? item.setField("date", newDate) : "";
-    await item.saveTx();
+    return item;
+    // await item.saveTx();
 }