+++ src/modules/rules/field-title.ts
@@ -35,7 +35,8 @@ async function titleCase2SentenceCase(item: Zotero.Item) {
     const title = item.getField("title", false, true) as string;
     const newTitle = toSentenceCase(title);
     item.setField("title", newTitle);
-    await item.saveTx();
+    return item;
+    // await item.saveTx();
 }
 
 /**
@@ -46,7 +47,8 @@ async function replaceGuillemetToBrackets(item: Zotero.Item) {
     const title = item.getField("title", false, true) as string;
     const newTitle = title.replace(/《/g, "〈").replace(/》/g, "〉");
     item.setField("title", newTitle);
-    await item.saveTx();
+    return item;
+    // await item.saveTx();
 }
 
 /**
@@ -58,7 +60,8 @@ async function replaceBracketsToGuillemet(item: Zotero.Item) {
     // const newTitle = title.replace(/<(?!\/|sub|sup|b|i)/g, "《").replace(/(?<!\/|sub|sup|b|i)>/g, "》");
     const newTitle = title.replace(/〈/g, "《").replace(/〉/g, "》");
     item.setField("title", newTitle);
-    await item.saveTx();
+    return item;
+    // await item.saveTx();
 }
 
 /**
@@ -70,7 +73,8 @@ async function replaceDoubleQuoteToSingleQuote(item: Zotero.Item) {
     const title = item.getField("title") as string;
     const newTitle = title.replace(/"/g, "'").replace(/“/g, "‘").replace(/”/g, "’");
     item.setField("title", newTitle);
-    await item.saveTx();
+    return item;
+    // await item.saveTx();
 }
 
 /**
@@ -82,5 +86,6 @@ async function replaceSingleQuoteToDoubleQuote(item: Zotero.Item) {
     const title = item.getField("title") as string;
     const newTitle = title.replace(/'/g, '"').replace(/‘/g, "“").replace(/’/g, "”");
     item.setField("title", newTitle);
-    await item.saveTx();
+    return item;
+    // await item.saveTx();
 }